requirements.txt
streamlit
spotipy
pandas
plotly
numpy
matplotlib
seaborn
python-dotenv